+++ BarRoulette/BarRoulette/OAuthConsumer/OAProblem.m
@@ -101,7 +101,7 @@ const NSString *token_not_renewable = @"token_not_renewable";
 	return NO;
 }
 
-- (int)code {
+- (NSUInteger)code {
 	return [[[self class] validProblems] indexOfObject:problem];
 }
